requirements.txt
invertransforms==0.2.1
numpy==1.22.0
opencv-python==4.5.5.62
Pillow==9.0.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
typing_extensions==4.0.1